helixmind/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9

groq==0.9.0

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9

chromadb==0.5.5
sentence-transformers==3.0.1

pypdf==4.3.1